requirements.txt
anyio==3.6.1
astroid==2.11.7
autopep8==1.7.0
bcrypt==4.0.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.6
dill==0.3.5.1
docopt==0.6.2
ecdsa==0.18.0
fastapi==0.79.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
passlib==1.7.4
pipreqs==0.4.11
platformdirs==2.5.2
psycopg2==2.9.3
pyasn1==0.4.8
pycodestyle==2.9.1
pydantic==1.9.2
PyJWT==2.5.0
pylint==2.14.5
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.40
starlette==0.19.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
typing_extensions==4.3.0
urllib3==1.26.11
uvicorn==0.18.2
wrapt==1.14.1
yarg==0.1.9
